Handle --help in the cli
The Windows docker builds expect that running the executable with `--help` will print out information, which isn't the case though since this plugin isn't using urfave/cli. For the moment just exit gracefully if `--help` is passed in.

This commit should be reverted when a proper docker pipleine is used on Windows rather than executing directly on the host.